Refactor annotation pipeline and update repository settings
This commit updated the annotation pipeline in 'pipeline.py' to remove bugs and close logical gaps. Code for 'authenticate', 'get_tasks', 'get_instances' and 'annotate' functions are updated to better handle tasks and instances. It also changes routes and parameters in 'repository-settings.json' to debug interactions with the server. In addition, '.gitignore' is updated to ignore log files.